Moved UsedToken component in its own separate file
